+++ soc-access/bookmarks/AdventureBookmarkSlots.cs
@@ -1,0 +1,10 @@
+namespace SongsOfConquestAccess.Bookmarks
+{
+    internal static class AdventureBookmarkSlots
+    {
+        public static readonly string[] All =
+        {
+            "1", "2", "3", "4", "5", "6", "7", "8", "9", "0"
+        };
+    }
+}